Split utils.py into more appropriate modules
