+++ docs/astralavista.rst
@@ -1,0 +1,8 @@
+.. Documents astralavista.py
+
+Astralavista
+============
+
+.. automodule:: astralavista
+    :members:
+    :undoc-members:
