+++ docs/error.rst
@@ -1,0 +1,8 @@
+.. Documents error.py
+
+Error
+=====
+
+.. automodule:: error
+    :members:
+    :undoc-members:
