+++ docs/latticejson.rst
@@ -1,0 +1,8 @@
+.. Documents latticejson.py
+
+LatticeJSON
+===========
+
+.. automodule:: latticejson
+    :members:
+    :undoc-members:
