+++ docs/nocelot.rst
@@ -1,0 +1,8 @@
+.. Documents nocelot.py
+
+NOcelot
+=======
+
+.. automodule:: nocelot
+    :members:
+    :undoc-members:
