--- docs/utils.rst
@@ -1,8 +1,0 @@
-.. Documents utils.py
-
-Utilities
-=========
-
-.. automodule:: utils
-    :members:
-    :undoc-members:
