+++ test/test_lattice_json.py
@@ -2,7 +2,7 @@ import os
 import test.ARESlatticeStage3v1_9 as ares
 
 from cheetah.accelerator import Segment
-from cheetah.utils import load_cheetah_model, save_cheetah_model
+from cheetah.latticejson import load_cheetah_model, save_cheetah_model
 
 cheetah_segment = Segment.from_ocelot(ares.cell, name="ARES_Segment")
